requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langsmith>=0.0.77

# Vector stores and embeddings
chromadb>=0.4.22
tiktoken>=0.5.2
sentence-transformers>=2.2.0

# Research APIs
semanticscholar>=0.3.0
arxiv>=2.0.0

# ML models (for Week 2)
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
setfit>=1.0.0
datasets>=2.14.0
scipy>=1.11.0

# Caching
redis>=5.0.0

# Document generation
python-pptx>=0.6.21
matplotlib>=3.7.0
Pillow>=10.0.0

# Additional utilities
aiohttp>=3.9.0
asyncio>=3.4.3
